Finished Exercise 4.21

+++ controllers/blogs.js
@@ -61,8 +61,25 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blogToDelete = await Blog.findById(request.params.id)
+  const blogCreatorUserId = blogToDelete.user.toString()
+
+  if (blogCreatorUserId === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(401).json({ error: 'you are not logged in as the user who saved this blog post' })
+  }
+
+  
 })
